move music player to own component
