+++ src/app/home/home.component.html
@@ -7,79 +7,7 @@
     <p>Explore our products ></p>
   </div>
 </div>
-<div class="container-fluid recordplayer-container">
-  <div class="vinyl-record" #recordplayer></div>
-  <div class="play-buttons">
+<app-music-player></app-music-player>
 
-    <audio id="heartbeat" #audio src="../../assets/audio/record0.mp3" preload="auto">
-      Your browser does not support the <code>audio</code> element.
-    </audio>
-
-    <h3>New Release</h3>
-    <ol #listofrecords>
-      <li>
-        <div class="music-card" [ngClass]="{'active': activeTrack == 0}" (click)="playRecord(0)">
-          <div class="flex-row">
-            1 Freaking Out (Remix)    <button class="play">Play / Pause</button>
-          </div>
-          <div class="track-progress flex-row">
-            <div class="total-track-duration">
-              <div class="percent-played track-zero" #percentplayed></div>
-            </div>
-          </div>
-        </div>
-      </li>
-      <li>
-        <div class="music-card" [ngClass]="{'active': activeTrack == 1}" (click)="playRecord(1)">
-          <div class="flex-row">
-            2 I Don't Care (Part I)<button class="play">Play / Pause</button>
-          </div>
-        <div class="track-progress flex-row">
-          <div class="total-track-duration">
-            <div class="percent-played track-one"></div>
-          </div>
-        </div>
-      </div>
-      </li>
-      <li>
-        <div class="music-card" [ngClass]="{'active': activeTrack == 2}" (click)="playRecord(2)">
-          <div>
-            3 I Don't Care (Part II)<button class="play">Play / Pause</button>
-          </div>
-          <div class="track-progress flex-row">
-            <div class="total-track-duration">
-              <div class="percent-played track-one"></div>
-            </div>
-          </div>
-        </div>
-      </li>
-      <li>
-        <div class="music-card" [ngClass]="{'active': activeTrack == 3}" (click)="playRecord(3)">
-          <div>  
-            4 Christmas Carols<button class="play">Play / Pause</button>
-          </div>
-          <div class="track-progress flex-row">
-            <div class="total-track-duration">
-              <div class="percent-played track-one"></div>
-            </div>
-          </div>
-        </div>
-      </li>
-      <li>
-        <div class="music-card" [ngClass]="{'active': activeTrack == 4}" (click)="playRecord(4)">
-          <div>
-            5 I Don't Care<button class="play">Play / Pause</button>
-          </div>
-          <div class="track-progress flex-row">
-            <div class="total-track-duration">
-              <div class="percent-played track-one"></div>
-            </div>
-          </div>
-        </div>
-      </li>
-    </ol>
-
-  </div>
-</div>
