+++ src/app/music-player/music-player.component.css
@@ -1,0 +1,67 @@
+
+.recordplayer-container{
+    /* padding: 56px 0px; */
+    background: rgb(59,59,93);
+    background: linear-gradient(167deg, rgba(59,59,93,1) 0%, rgba(16,26,39,1) 100%); 
+    display: flex;
+    justify-content: center;
+  }
+  
+  .vinyl-record {
+    background-image: url(../../assets/images/vinyl_record.png);
+    background-repeat: no-repeat;
+    width: 250px;
+    height: 250px;
+    background-size: 250px 250px;
+  }
+  
+  @keyframes spinrecord {
+    /* 0%   {background-color:red; left:0px; top:0px;} */
+    25%  {transform: rotate(90deg);}
+    50%  {transform: rotate(180deg);}
+    75%  {transform: rotate(270deg);}
+    100% {transform: rotate(360deg);}
+  }
+  
+  .vinyl-record.spinning {
+    animation: 2s linear infinite spinrecord;
+  }
+  
+  .play-buttons {
+    text-align: center;
+  }
+  
+  ol {
+    margin: 0;
+    text-align: left;
+  }
+  
+  ol li {
+    padding: 8px;
+  }
+  
+  .music-card {
+    background-color: black;
+    color: white;
+    padding: 8px;
+  }
+  
+  .music-card .flex-row.track-progress {
+    display: none;
+  } 
+  
+  .music-card.active .flex-row.track-progress {
+    display: flex;
+  } 
+  
+  .total-track-duration {
+    background-color: lightgrey;
+    width: 100%;
+    height: 8px;
+  }
+  
+  .percent-played {
+    background-color: violet;
+    width: 0%;
+    height: 8px;
+  }
